style: moves components to separate files
