+++ app/[locale]/search.tsx
@@ -1,0 +1,7 @@
+import type { FC } from 'react';
+
+const SearchPage: FC = () => {
+  return <div>Search page</div>;
+};
+
+export default SearchPage;
