+++ components/SearchBox/components/EmptyState.tsx
@@ -1,0 +1,7 @@
+import type { FC } from 'react';
+
+import styles from './index.module.css';
+
+export const EmptyState: FC = () => {
+  return <div className={styles.emptyStateContainer}>Search something...</div>;
+};
