+++ components/SearchBox/components/NoResults.tsx
@@ -1,0 +1,16 @@
+import type { FC } from 'react';
+
+import styles from './index.module.css';
+
+type NoResultsProps = {
+  searchTerm: string;
+};
+
+export const NoResults: FC<NoResultsProps> = props => {
+  return (
+    <div className={styles.noResultsContainer}>
+      No results for&nbsp;
+      <span className={styles.noResultsTerm}>{`"${props.searchTerm}"`}</span>.
+    </div>
+  );
+};
