+++ components/SearchBox/index.module.css
@@ -27,142 +27,3 @@
     h-5
     w-5;
 }
-
-.searchBoxModalContainer {
-  @apply fixed
-    inset-0
-    z-50
-    flex
-    items-center
-    justify-center
-    bg-neutral-900
-    bg-opacity-90
-    dark:bg-neutral-900
-    dark:bg-opacity-90;
-}
-
-.searchBoxModalPanel {
-  @apply fixed
-    top-60
-    w-full
-    max-w-3xl
-    rounded-xl
-    bg-neutral-100
-    p-2
-    shadow-lg
-    dark:bg-neutral-950;
-}
-
-.searchBoxInnerPanel {
-  @apply p-2
-  text-neutral-800
-  dark:text-neutral-400;
-}
-
-.searchBoxMagnifyingGlassIcon {
-  @apply absolute top-[10px] h-6 w-6;
-}
-
-.searchBoxInputContainer {
-  @apply relative px-2;
-}
-
-.searchBoxInput {
-  @apply w-full
-    border-b
-    border-neutral-300
-    bg-transparent
-    py-2
-    pl-8
-    pr-4
-    focus:outline-none
-  dark:border-neutral-900
-  dark:text-neutral-300
-  dark:placeholder-neutral-300;
-}
-
-.fulltextResultsContainer {
-  @apply h-80 overflow-scroll;
-}
-
-.fulltextSearchResult {
-  @apply flex
-    flex-col
-    rounded-md
-    p-2
-    text-left
-    text-sm
-    hover:bg-neutral-300
-    dark:hover:bg-neutral-900;
-}
-
-.fulltextSearchResultTitle {
-  @apply text-neutral-700
-    dark:text-neutral-300;
-}
-
-.fulltextSearchResultBreadcrumb {
-  @apply mt-1 text-xs capitalize text-neutral-800
-    dark:text-neutral-600;
-}
-
-.fulltextSearchSections {
-  @apply mb-1
-    mt-2
-    flex
-    gap-2
-    p-2
-    text-xs
-    font-semibold
-    text-neutral-700
-    dark:text-neutral-600;
-}
-
-.fulltextSearchSection {
-  @apply rounded-lg
-    border-b
-    border-transparent
-    px-2
-    py-1
-    capitalize
-    hover:bg-neutral-200
-    dark:border-neutral-900
-    dark:border-b-transparent
-    dark:hover:bg-neutral-900;
-}
-
-.fulltextSearchSectionSelected {
-  @apply rounded-b-none border-neutral-700 text-neutral-900 dark:border-neutral-700 dark:text-neutral-300;
-}
-
-.fulltextSearchSectionCount {
-  @apply ml-1 text-neutral-500 dark:text-neutral-800;
-}
-
-.seeAllFulltextSearchResults {
-  @apply m-auto
-    mb-2
-    mt-4
-    w-full
-    text-center
-    text-sm
-  text-neutral-700
-    hover:underline
-  dark:text-neutral-600;
-}
-
-.poweredBy {
-  @apply absolute
-    -bottom-8
-    left-0
-    flex
-    w-full
-    items-center
-    justify-center
-    text-xs
-    text-neutral-200;
-}
-
-.poweredByLogo {
-  @apply ml-2 w-16;
-}
