+++ components/SearchBox/lib/utils.ts
@@ -1,0 +1,8 @@
+export function pathToBreadcrumbs(path: string) {
+  return path
+    .replace(/#.+$/, '')
+    .split('/')
+    .slice(0, -1)
+    .map(element => element.replaceAll('-', ' '))
+    .filter(Boolean);
+}
